Refactor to only need one test file
